requirements.txt
# Core Flask
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Swagger/API Documentation
flasgger==0.9.7.1
flask-swagger-ui==4.11.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Playwright & Testing
playwright==1.40.0
behave==1.2.6

# Google Search
google-search-results==2.4.2

# Utilities
pyyaml==6.0.1
requests==2.31.0
jinja2==3.1.2
pyautogui==0.9.54

